# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: VPS Website Manager API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /sites:
    get:
      operationId: listSites
      tags: [sites]
      summary: List all deployed websites
      responses:
        "200":
          description: List of sites
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Site"
    post:
      operationId: createSite
      tags: [sites]
      summary: Deploy a new website
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateSiteBody"
      responses:
        "201":
          description: Created site
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Site"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /sites/{id}:
    get:
      operationId: getSite
      tags: [sites]
      summary: Get a specific site
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Site details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Site"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updateSite
      tags: [sites]
      summary: Update site configuration
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateSiteBody"
      responses:
        "200":
          description: Updated site
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Site"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deleteSite
      tags: [sites]
      summary: Remove a deployed site
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /sites/{id}/deploy:
    post:
      operationId: deploySite
      tags: [sites]
      summary: Trigger a deployment/re-deployment of a site
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deploy result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CommandResult"
  /sites/{id}/ssl:
    post:
      operationId: installSsl
      tags: [sites]
      summary: Install SSL certificate using Let's Encrypt
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: SSL install result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CommandResult"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...
    CommandResult:
      type: object
      properties:
        success:
          type: boolean
        output:
          type: string
        error:
          type: ["string", "null"]
      required:
        - success
        - output
...
    Site:
      type: object
      properties:
        id:
          type: integer
        serverId:
          type: integer
        name:
          type: string
        domain:
          type: string
        repoUrl:
          type: ["string", "null"]
        repoToken:
          type: ["string", "null"]
        deployPath:
          type: string
        buildCommand:
          type: ["string", "null"]
        siteType:
          type: string
          enum: [static, nodejs, php, python]
        status:
          type: string
          enum: [deploying, active, failed, stopped]
        sslInstalled:
          type: boolean
        autoSync:
          type: boolean
        lastDeployedAt:
          type: ["string", "null"]
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - serverId
        - name
        - domain
        - deployPath
        - siteType
        - status
        - sslInstalled
        - autoSync
        - createdAt
        - updatedAt
    CreateSiteBody:
      type: object
      properties:
        serverId:
          type: integer
        name:
          type: string
        domain:
          type: string
        repoUrl:
          type: ["string", "null"]
        repoToken:
          type: ["string", "null"]
        deployPath:
          type: string
        buildCommand:
          type: ["string", "null"]
        siteType:
          type: string
          enum: [static, nodejs, php, python]
        autoSync:
          type: boolean
      required:
        - serverId
        - name
        - domain
        - deployPath
        - siteType
    UpdateSiteBody:
      type: object
      properties:
        name:
          type: string
        domain:
          type: string
        repoUrl:
          type: ["string", "null"]
        repoToken:
          type: ["string", "null"]
        buildCommand:
          type: ["string", "null"]
        siteType:
          type: string
          enum: [static, nodejs, php, python]
        autoSync:
          type: boolean
...