# lib/api-zod/src/generated/types/createSiteBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * VPS Website Manager API
 * OpenAPI spec version: 0.1.0
 */
import type { CreateSiteBodySiteType } from "./createSiteBodySiteType";

export interface CreateSiteBody {
  serverId: number;
  name: string;
  domain: string;
  /** @nullable */
  repoUrl?: string | null;
  /** @nullable */
  repoToken?: string | null;
  deployPath: string;
  /** @nullable */
  buildCommand?: string | null;
  siteType: CreateSiteBodySiteType;
  autoSync?: boolean;
}