# lib/api-zod/src/generated/types/site.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * VPS Website Manager API
 * OpenAPI spec version: 0.1.0
 */
import type { SiteSiteType } from "./siteSiteType";
import type { SiteStatus } from "./siteStatus";

export interface Site {
  id: number;
  serverId: number;
  name: string;
  domain: string;
  /** @nullable */
  repoUrl?: string | null;
  /** @nullable */
  repoToken?: string | null;
  deployPath: string;
  /** @nullable */
  buildCommand?: string | null;
  siteType: SiteSiteType;
  status: SiteStatus;
  sslInstalled: boolean;
  autoSync: boolean;
  /** @nullable */
  lastDeployedAt?: string | null;
  createdAt: string;
  updatedAt: string;
}