# lib/api-zod/src/generated/types/updateSiteBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * VPS Website Manager API
 * OpenAPI spec version: 0.1.0
 */
import type { UpdateSiteBodySiteType } from "./updateSiteBodySiteType";

export interface UpdateSiteBody {
  name?: string;
  domain?: string;
  /** @nullable */
  repoUrl?: string | null;
  /** @nullable */
  repoToken?: string | null;
  /** @nullable */
  buildCommand?: string | null;
  siteType?: UpdateSiteBodySiteType;
  autoSync?: boolean;
}